chore: fix lint error on req.user access and restructure
