+++ packages/backend/src/SchedulerApp.ts
@@ -13,7 +13,7 @@ import Logger from './logging/logger';
 import { ModelProviderMap, ModelRepository } from './models/ModelRepository';
 import PrometheusMetrics from './prometheus';
 import { SchedulerWorker } from './scheduler/SchedulerWorker';
-import { IGNORE_ERRORS } from './sentry';
+import { IGNORE_ERRORS } from './sentry/init';
 import {
     OperationContext,
     ServiceProviderMap,
